// plugins/dsh-pet/src/shared/types.ts
export interface Rect {
  x: number;
  y: number;
  width: number;
  height: number;
}